sistema de usuarios funcionando

+++ views.py
@@ -1,35 +1,59 @@
 from flask import Blueprint, request, redirect, render_template, url_for
 from flask.ext.mongoengine.wtf import model_form
+from flask.ext.login import (current_user, login_required, login_user, logout_user, confirm_login, fresh_login_required)
 from flask.views import MethodView
 
 from voteaqui.models import *
+from datetime import datetime
 
 polls = Blueprint('polls', __name__, template_folder='templates')
 
 class ListView(MethodView):
+	@login_required
 	def get(self):
-		polls = Poll.objects.all()
+		polls = []
+		user = User.objects.get(id=current_user.id)
+		can_vote = True
+		date = datetime.now()
+		for poll in Poll.objects:
+			if poll.expiration_date > date:
+				polls.append(poll)
+
 		return render_template('polls/list.html', polls=polls)
 
 
 class DetailView(MethodView):
-	form = model_form(Comment)
+	form = model_form(Comment, exclude=['author'])
 
 	def get_context(self, poll_id):
 		poll = Poll.objects.get_or_404(id=poll_id)
 		form = self.form(request.form)
+		user = User.objects.get(id=current_user.id)
+
+		can_vote = True
+		'''
+		if poll.choices:
+			for choice in poll.choices:
+				for choice_user in choice.users:
+					if choice_user == user:
+						can_vote = False
+		'''
 
 		context = {
 			"poll": poll,
-			"form": form
+			"form": form,
+			"user": user,
+			"can_vote": can_vote
 		}
 
 		return context
 
+	@login_required
 	def get(self, poll_id):
 		context = self.get_context(poll_id)
 		return render_template('polls/detail.html', **context)
 
+	@login_required
 	def post(self, poll_id):
 		context = self.get_context(poll_id)
 		form = context.get('form')
@@ -37,16 +61,33 @@ class DetailView(MethodView):
 		if form.validate():
 			comment = Comment()
 			form.populate_obj(comment)
+			comment.author = context.get('user')
+			comment.save()
 
 			poll = context.get('poll')
 			poll.comments.append(comment)
 			poll.save()
+
+			user = context.get('user')
+			user.comments.append(comment)
+			user.save()
 			return redirect(url_for('polls.detail', poll_id=poll_id))
 		return render_template('polls/detail.html', **context)
 
 
+class VoteView(MethodView):
+	@login_required
+	def get(self, poll_id, choice_id):
+		choice = Choice.objects.get(id=choice_id)
+		choice.plus_one_vote()
+		user = User.objects.get(id=current_user.id)
+		choice.users.append(user)
+		choice.save()
+		return redirect(url_for('polls.detail', poll_id=poll_id))
+
+
 class PollCreateView(MethodView):
-	form = model_form(Poll, exclude=['created_at', 'comments', 'choices'])
+	form = model_form(Poll, exclude=['created_at', 'comments', 'choices', 'author', 'expiration_date'])
 
 	def get_context(self):
 		poll = Poll()
@@ -59,25 +100,34 @@ class PollCreateView(MethodView):
 
 		return context
 
+	@login_required
 	def get(self):
 		context = self.get_context()
 		return render_template('polls/poll_form.html', **context)
 
+	@login_required
 	def post(self):
 		context = self.get_context()
 		form = context.get('form')
-		print request.form['expiration_date']
+		date = datetime(int(request.form['expiration_date'][:4]), int(request.form['expiration_date'][5:7]), int(request.form['expiration_date'][8:10]), 0, 0, 0)
+
 		if form.validate():
 			poll = context.get('poll')
 			form.populate_obj(poll)
+			poll.expiration_date = date
+			user = User.objects.get(id=current_user.id)
+			poll.author = user
 			poll.save()
+
+			user.polls.append(poll)
+			user.save()
 
 			return redirect(url_for('polls.list'))
 		return render_template('polls/poll_form.html', **context)
 
 
 class ChoiceCreateView(MethodView):
-	form = model_form(Choice, exclude=['votes'])
+	form = model_form(Choice, exclude=['votes', 'users'])
 
 	def get_context(self, poll_id):
 		poll = Poll.objects.get_or_404(id=poll_id)
@@ -90,10 +140,12 @@ class ChoiceCreateView(MethodView):
 
 		return context
 
+	@login_required
 	def get(self, poll_id):
 		context = self.get_context(poll_id)
 		return render_template('polls/choice_form.html', **context)
 
+	@login_required
 	def post(self, poll_id):
 		context = self.get_context(poll_id)
 		form = context.get('form')
@@ -101,6 +153,7 @@ class ChoiceCreateView(MethodView):
 		if form.validate():
 			choice = Choice()
 			form.populate_obj(choice)
+			choice.save()
 
 			poll = context.get('poll')
 			poll.choices.append(choice)
@@ -110,24 +163,40 @@ class ChoiceCreateView(MethodView):
 
 
 class PollDeleteView(MethodView):
+	@login_required
 	def get(self, poll_id):
 		poll = Poll.objects.get_or_404(id=poll_id)
-		poll.delete()
+		user = User.objects.get(id=current_user.id)
+		if poll.author == user:
+			poll.delete()
 		return redirect(url_for('polls.list'))
 
 
 class ChoiceDeleteView(MethodView):
+	@login_required
 	def get(self, poll_id, choice_id):
+		poll = Poll.objects.get(id=poll_id)
 		choice = Choice.objects.get_or_404(id=choice_id)
-		choice.delete()
-		return redirect(url_for('polls.detail'), poll_id=poll_id)
+		user = User.objects.get(id=current_user.id)
+		if poll.author == user:
+			poll.choices.remove(choice)
+			poll.save()
+			choice.delete()
+		return redirect(url_for('polls.detail', poll_id=poll_id))
 
 
 class CommentDeleteView(MethodView):
+	@login_required
 	def get(self, poll_id, comment_id):
 		comment = Comment.objects.get_or_404(id=comment_id)
-		comment.delete()
-		return redirect(url_for('polls.detail'), poll_id=poll_id)
+		user = User.objects.get(id=current_user.id)
+		if comment.author == user:
+			poll.comments.remove(comment)
+			user.comments.remove(comment)
+			poll.save()
+			user.save()
+			comment.delete()
+		return redirect(url_for('polls.detail', poll_id=poll_id))
 
 
 polls.add_url_rule('/', view_func=ListView.as_view('list'))
@@ -135,5 +204,6 @@ polls.add_url_rule('/<poll_id>/', view_func=DetailView.as_view('detail'))
 polls.add_url_rule('/create/', view_func=PollCreateView.as_view('create'))
 polls.add_url_rule('/create/<poll_id>/', view_func=ChoiceCreateView.as_view('create_choice'))
 polls.add_url_rule('/delete/<poll_id>/', view_func=PollDeleteView.as_view('delete'))
-polls.add_url_rule('/delete_choice/<poll_id>/<choice_id>/', view_func=PollDeleteView.as_view('delete_choice'))
-polls.add_url_rule('/delete/<poll_id>/<comment_id>/', view_func=PollDeleteView.as_view('delete_comment'))
+polls.add_url_rule('/delete_choice/<poll_id>/<choice_id>/', view_func=ChoiceDeleteView.as_view('delete_choice'))
+polls.add_url_rule('/delete/<poll_id>/<comment_id>/', view_func=CommentDeleteView.as_view('delete_comment'))
+polls.add_url_rule('/vote/<poll_id>/<choice_id>/', view_func=VoteView.as_view('vote'))
